# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Behavioral Home Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vital-signs:
    get:
      operationId: listVitalSigns
      tags: [medication-safety]
      summary: List vital signs records
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: administrationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Vital signs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VitalSigns"
    post:
      operationId: createVitalSigns
      tags: [medication-safety]
      summary: Record vital signs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVitalSignsBody"
      responses:
        "201":
          description: Vital signs recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VitalSigns"
  /medication-side-effects:
    get:
      operationId: listMedicationSideEffects
      tags: [medication-safety]
      summary: List medication side effects
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: medicationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Side effects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicationSideEffect"
    post:
      operationId: createMedicationSideEffect
      tags: [medication-safety]
      summary: Report a medication side effect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicationSideEffectBody"
      responses:
        "201":
          description: Side effect reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationSideEffect"
  /medication-refusals:
    get:
      operationId: listMedicationRefusals
      tags: [medication-safety]
      summary: List medication refusals
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Refusals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicationRefusal"
    post:
      operationId: createMedicationRefusal
      tags: [medication-safety]
      summary: Record a medication refusal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicationRefusalBody"
      responses:
        "201":
          description: Refusal recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationRefusal"
  /medication-audit-log:
    get:
      operationId: listMedicationAuditLog
      tags: [medication-safety]
      summary: List medication audit trail
      parameters:
        - name: entityType
          in: query
          schema:
            type: string
        - name: entityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicationAuditLogEntry"
  /medication-administrations/{id}/prn-followup:
    put:
      operationId: updatePrnFollowup
      tags: [medication-safety]
      summary: Record PRN effectiveness follow-up
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrnFollowupBody"
      responses:
        "200":
          description: PRN follow-up recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationAdministration"
...
components:
  schemas:
...
    MedicationAdministration:
      type: object
      properties:
        id:
          type: integer
        medicationId:
          type: integer
        medicationName:
          type: ["string", "null"]
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        administeredAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [given, missed, refused, held]
        notes:
          type: ["string", "null"]
        scheduledTime:
          type: ["string", "null"]
          format: date-time
        windowStart:
          type: ["string", "null"]
          format: date-time
        windowEnd:
          type: ["string", "null"]
          format: date-time
        prnReason:
          type: ["string", "null"]
        prnEffectiveness:
          type: ["string", "null"]
        prnFollowUpAt:
          type: ["string", "null"]
          format: date-time
        prnFollowUpNotes:
          type: ["string", "null"]
        barcodeScanVerified:
          type: boolean
        witnessStaffId:
          type: ["integer", "null"]
        relatedIncidentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, medicationId, medicationName, patientId, patientName, staffId, staffName, administeredAt, status, notes, scheduledTime, windowStart, windowEnd, prnReason, prnEffectiveness, prnFollowUpAt, prnFollowUpNotes, barcodeScanVerified, witnessStaffId, relatedIncidentId, createdAt]
...
    VitalSigns:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        staffId:
          type: integer
        staffName:
          type: string
        administrationId:
          type: ["integer", "null"]
        systolicBp:
          type: ["integer", "null"]
        diastolicBp:
          type: ["integer", "null"]
        heartRate:
          type: ["integer", "null"]
        temperature:
          type: ["number", "null"]
        respiratoryRate:
          type: ["integer", "null"]
        oxygenSaturation:
          type: ["integer", "null"]
        painLevel:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        recordedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, patientId, patientName, staffId, staffName, recordedAt, createdAt]
    CreateVitalSignsBody:
      type: object
      properties:
        patientId:
          type: integer
        staffId:
          type: integer
        administrationId:
          type: ["integer", "null"]
        systolicBp:
          type: ["integer", "null"]
        diastolicBp:
          type: ["integer", "null"]
        heartRate:
          type: ["integer", "null"]
        temperature:
          type: ["number", "null"]
        respiratoryRate:
          type: ["integer", "null"]
        oxygenSaturation:
          type: ["integer", "null"]
        painLevel:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        recordedAt:
          type: string
          format: date-time
      required: [patientId, staffId]
    MedicationSideEffect:
      type: object
      properties:
        id:
          type: integer
        medicationId:
          type: integer
        medicationName:
          type: string
        patientId:
          type: integer
        patientName:
          type: string
        staffId:
          type: integer
        staffName:
          type: string
        administrationId:
          type: ["integer", "null"]
        sideEffect:
          type: string
        severity:
          type: string
          enum: [mild, moderate, severe, life_threatening]
        onsetTime:
          type: string
          format: date-time
        resolved:
          type: boolean
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, medicationId, medicationName, patientId, patientName, staffId, staffName, sideEffect, severity, onsetTime, resolved, createdAt]
    CreateMedicationSideEffectBody:
      type: object
      properties:
        medicationId:
          type: integer
        patientId:
          type: integer
        staffId:
          type: integer
        administrationId:
          type: ["integer", "null"]
        sideEffect:
          type: string
        severity:
          type: string
          enum: [mild, moderate, severe, life_threatening]
        onsetTime:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
      required: [medicationId, patientId, staffId, sideEffect, severity]
    MedicationRefusal:
      type: object
      properties:
        id:
          type: integer
        medicationId:
          type: integer
        medicationName:
          type: string
        patientId:
          type: integer
        patientName:
          type: string
        staffId:
          type: integer
        staffName:
          type: string
        scheduledTime:
          type: ["string", "null"]
          format: date-time
        reason:
          type: string
        physicianNotified:
          type: boolean
        physicianNotifiedAt:
          type: ["string", "null"]
          format: date-time
        physicianName:
          type: ["string", "null"]
        followUpAction:
          type: ["string", "null"]
        followUpNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, medicationId, medicationName, patientId, patientName, staffId, staffName, reason, physicianNotified, createdAt]
    CreateMedicationRefusalBody:
      type: object
      properties:
        medicationId:
          type: integer
        patientId:
          type: integer
        staffId:
          type: integer
        scheduledTime:
          type: ["string", "null"]
          format: date-time
        reason:
          type: string
        physicianNotified:
          type: boolean
        physicianName:
          type: ["string", "null"]
        followUpAction:
          type: ["string", "null"]
        followUpNotes:
          type: ["string", "null"]
      required: [medicationId, patientId, staffId, reason]
    MedicationAuditLogEntry:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        action:
          type: string
        performedBy:
          type: integer
        performedByName:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        previousValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, entityType, entityId, action, performedBy, createdAt]
    PrnFollowupBody:
      type: object
      properties:
        prnEffectivenessScore:
          type: integer
          minimum: 1
          maximum: 10
        prnFollowUpNotes:
          type: string
        prnEffectiveness:
          type: string
          enum: [effective, partially_effective, ineffective]
      required: [prnEffectivenessScore, prnEffectiveness]
...